refactor to keep up with simulation measurement messages by averaging zary values to drain any pending queue items and only do the full estimate processing when the queue is empty;  added a thread-safe concurrent/shared queue with the SELoopConsumer class only putting new measurement messages on the queue and the new SELoopWorker class popping messages from the and doing all the processing in a separate (main) thread
